Move docker keyring lookup test to pkg/credentailprovider
Also remove unused image tests in docker_test.go
